MainModel/requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-generativeai==0.3.1
python-dotenv==0.19.0
Werkzeug==2.0.1
click==7.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1 